unit testing

+++ week-one/src/main/java/com/qa/springdemo/service/CatService.java
@@ -14,5 +14,5 @@ public interface CatService {
 
 	Cat remove(int id);
 
-	Cat update(int id, String name, Boolean evil, Boolean hasWhiskers, Integer length);
+	Cat update(long id, String name, Boolean evil, Boolean hasWhiskers, Integer length);
 }
